Allow `dnu build` and `dnu pack` to work with globbing patterns and build all projects at once. This is a big perf improvement because each project is compiled only once (caches are shared). Change is 100% backward compatible with the old one-project-one-build.

The more projects you have, the better the results are. MVC has 77 projects in `src` and `test`  combined.

Perf results summary:

```
src:      60s -> 15s
test:     436s -> 106s
src+test: 493s -> 111s
```

Full perf results:

```
Measure-Command { Get-ChildItem -path 'src' -filter project.json -Recurse | Select -expand FullName | Split-Path | %{ dnu build $_ }} | Select -expand TotalSeconds
60.3610104
Measure-Command { dnu build src/** } | Select -expand TotalSeconds
15.2666252
Measure-Command { Get-ChildItem -path 'test' -filter project.json -Recurse | Select -expand FullName | Split-Path | %{ dnu build $_ }} | Select -expand TotalSeconds
436.9794821
Measure-Command { dnu build  test/** } | Select -expand TotalSeconds
103.2918838
Measure-Command { Get-ChildItem -path 'src','test' -filter project.json -Recurse | Select -expand FullName | Split-Path | %{ dnu build $_ }} | Select -expand TotalSeconds
493.5432721
Measure-Command { dnu build  src/** test/** } | Select -expand TotalSeconds
111.036362
```

Please review: @davidfowl @ChengTian