@pranavkm 2015-06-12
Change looks pretty solid :shipit: 

As a follow up, we should figure out how to make it easy to run tests against packages - since we don't build test projects, simply run them. This would avoid having to recompile src.