@victorhurdugaci 2015-06-13
@pranavkm yup, that's on my list. during the lite build I'll generate all the packages (src + test) and then in the second pass run tests using the packages. There might be a few repos that don't support this, tho (MusicStore) is one of them